+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.?.?
+
+* BC break: Result() end event does not emit data anymore
+
 ## 0.1.0 (2017-07-25)
 
 * First tagged release